até a pagina 4 dos iniciantes
Adição e organização dos codigos

+++ URI 1040 a 1099/1060 - Números Positivos.py
@@ -1,6 +1,6 @@
 positivo = 0
 for x in range(6):
-    n = int(input())
+    n = float()(input())
     if n >= 0:
         positivo += 1
 print('{} valores positivos'.format(positivo))
